basically make everything work
