+++ include/pub_map.hpp
@@ -23,6 +23,54 @@
 #include <std_msgs/msg/string.hpp>
 
 
+namespace util
+{
+    template<typename ros_T, typename primitive_T>
+    inline constexpr ros_T& to_ros(primitive_T& v)
+    {
+        static_assert(
+            std::is_same<typename ros_T::_data_type, primitive_T>::value &&
+            sizeof(ros_T) == sizeof(primitive_T) &&
+            alignof(ros_T) == alignof(primitive_T) );
+
+        return reinterpret_cast<ros_T&>(v);
+    }
+
+    inline constexpr std_msgs::msg::Bool& to_ros(bool& v)
+    {
+        return reinterpret_cast<std_msgs::msg::Bool&>(v);
+    }
+    inline constexpr std_msgs::msg::Int64& to_ros(int64_t& v)
+    {
+        return reinterpret_cast<std_msgs::msg::Int64&>(v);
+    }
+    inline constexpr std_msgs::msg::Float64& to_ros(double& v)
+    {
+        return reinterpret_cast<std_msgs::msg::Float64&>(v);
+    }
+
+    template<typename ros_T, typename primitive_T>
+    inline ros_T to_ros_val(primitive_T v)
+    {
+        static_assert(std::is_same<typename ros_T::_data_type, primitive_T>::value);
+
+        return ros_T{}.set__data(v);
+    }
+
+    inline std_msgs::msg::Bool to_ros_val(bool v)
+    {
+        return std_msgs::msg::Bool{}.set__data(v);
+    }
+    inline std_msgs::msg::Int64 to_ros_val(int64_t v)
+    {
+        return std_msgs::msg::Int64{}.set__data(v);
+    }
+    inline std_msgs::msg::Float64 to_ros_val(double v)
+    {
+        return std_msgs::msg::Float64{}.set__data(v);
+    }
+};
+
 // template<typename...>
 // struct any_same : std::true_type {};
 // template<typename T, typename t>
