+++ src/engine/localization.cpp
@@ -12,11 +12,6 @@
 
 #include <boost/algorithm/string.hpp>
 
-// #ifdef USE_CV_BRIDGE_HPP
-// #include <cv_bridge/cv_bridge.hpp>
-// #else
-// #include <cv_bridge/cv_bridge.h>
-// #endif
 #include <pcl_conversions/pcl_conversions.h>
 
 #include <geometry_msgs/msg/transform_stamped.hpp>
@@ -24,8 +19,6 @@
 
 #include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 #include <tf2_sensor_msgs/tf2_sensor_msgs.hpp>
-
-// #include <opencv2/imgproc.hpp>
 
 
 using namespace util::geom::cvt::ops;
@@ -36,16 +29,12 @@ PerceptionNode::PerceptionNode() :
     tf_buffer{ std::make_shared<rclcpp::Clock>(RCL_ROS_TIME) },
     tf_listener{ tf_buffer },
     tf_broadcaster{ *this },
-    // img_transport{ std::shared_ptr<PerceptionNode>(this, [](auto*){}) },
     mt_callback_group{ this->create_callback_group(rclcpp::CallbackGroupType::Reentrant) },
     metrics_pub{ this, "/perception_debug/", 1 },
     pose_pub{ this, "/perception_debug/", 1 },
     lidar_odom{ this },
-    // tag_detection{ this },
     trajectory_filter{}
 {
-    // RCLCPP_INFO(this->get_logger(), "CONSTRUCTOR INIT");
-
     this->getParams();
     this->initMetrics();
 
@@ -63,6 +52,9 @@ PerceptionNode::PerceptionNode() :
     rclcpp::SubscriptionOptions ops{};
     ops.callback_group = this->mt_callback_group;
 
+    this->detections_sub = this->create_subscription<cardinal_perception::msg::TagsDetection>(
+        "tags_detections", rclcpp::SensorDataQoS{},
+        [this](const cardinal_perception::msg::TagsDetection::ConstSharedPtr& det){ this->detection_callback(det); }, ops);
     this->scan_sub = this->create_subscription<sensor_msgs::msg::PointCloud2>(
         scan_topic, rclcpp::SensorDataQoS{},
         [this](const sensor_msgs::msg::PointCloud2::ConstSharedPtr& scan){ this->scan_callback(scan); }, ops);
@@ -70,232 +62,21 @@ PerceptionNode::PerceptionNode() :
         imu_topic, rclcpp::SensorDataQoS{},
         [this](const sensor_msgs::msg::Imu::SharedPtr imu){ this->imu_callback(imu); }, ops);
 
-    // std::vector<std::string> img_topics, info_topics;
-    // util::declare_param(this, "img_topics", img_topics, {});
-    // util::declare_param(this, "info_topics", info_topics, {});
-
-    // const size_t n_img = img_topics.size(), n_info = info_topics.size();
-    // if(n_img > 0 && n_img <= n_info)
-    // {
-    //     this->camera_subs.resize(n_img);    // DO NOT CALL RESERVE AND SUBSEQUENTLY PUSH BACK INSTANCES!!!
-    //     for(size_t i = 0; i < n_img; i++)
-    //     {
-    //         this->camera_subs[i].initialize(this, img_topics[i], info_topics[i]);
-    //     }
-    // }
-
-    // this->debug_img_pub = this->img_transport.advertise("debug_img", rmw_qos_profile_sensor_data.depth);
     this->filtered_scan_pub = this->create_publisher<sensor_msgs::msg::PointCloud2>("filtered_scan", rclcpp::SensorDataQoS{});
+#if USE_GTSAM_PGO > 0
     this->path_pub = this->create_publisher<nav_msgs::msg::Path>("path", rclcpp::SensorDataQoS{});
-
-    // double trajectory_filter_params[6];
-    // util::declare_param(this, "trajectory_filter.window_duration", trajectory_filter_params[0], 0.5);
-    // util::declare_param(this, "trajectory_filter.min_sample_duration", trajectory_filter_params[1], 0.3);
-    // util::declare_param(this, "trajectory_filter.thresh.avg_linear_error", trajectory_filter_params[2], 2e-2);
-    // util::declare_param(this, "trajectory_filter.thresh.avg_angular_error", trajectory_filter_params[3], 5e-2);
-    // util::declare_param(this, "trajectory_filter.thresh.min_linear_variance", trajectory_filter_params[4], 1e-5);
-    // util::declare_param(this, "trajectory_filter.thresh.min_angular_variance", trajectory_filter_params[5], 1e-5);
-    // this->trajectory_filter.applyParams(
-    //     trajectory_filter_params[0],
-    //     trajectory_filter_params[1],
-    //     trajectory_filter_params[2],
-    //     trajectory_filter_params[3],
-    //     trajectory_filter_params[4],
-    //     trajectory_filter_params[5] );
-
-    // RCLCPP_INFO(this->get_logger(), "CONSTRUCTOR EXIT");
+#endif
 }
-
-
-// void PerceptionNode::CameraSubscriber::initialize(
-//     PerceptionNode* inst,
-//     const std::string& img_topic,
-//     const std::string& info_topic)
-// {
-//     if(!inst) return;
-//     this->pnode = inst;
-
-//     rclcpp::SubscriptionOptions ops{};
-//     ops.callback_group = this->pnode->mt_callback_group;
-
-//     this->image_sub = this->pnode->img_transport.subscribe(
-//         img_topic, rmw_qos_profile_sensor_data.depth,
-//         [this](const image_transport::ImageTransport::ImageConstPtr& img){ this->img_callback(img); },
-//         image_transport::ImageTransport::VoidPtr(), nullptr, ops);
-//     this->info_sub = this->pnode->create_subscription<sensor_msgs::msg::CameraInfo>(
-//         info_topic, rclcpp::SensorDataQoS{},
-//         [this](const image_transport::ImageTransport::CameraInfoConstPtr& info){ this->info_callback(info); }, ops);
-// }
-// PerceptionNode::CameraSubscriber::CameraSubscriber(
-//     const CameraSubscriber& ref
-// ) :
-//     pnode{ ref.pnode },
-//     dbg_frame{ ref.dbg_frame },
-//     calibration{ ref.calibration.clone() },
-//     distortion{ ref.distortion.clone() },
-//     valid_calib{ ref.valid_calib.load() }
-// {}
-
-// void PerceptionNode::CameraSubscriber::img_callback(const sensor_msgs::msg::Image::ConstSharedPtr& img)
-// {
-//     // RCLCPP_INFO(this->pnode->get_logger(), "IMG_CALLBACK -- %s", (std::stringstream{} << std::this_thread::get_id()).str().c_str());
-//     auto _start = std::chrono::system_clock::now();
-
-//     std::vector<TagDetection::Ptr> detections;
-//     this->pnode->tag_detection.processImg(img, *this, detections);
-//     this->pnode->state.any_new_frames = true;
-
-//     // if(detections.size() > 0)
-//     // {
-//     //     this->pnode->trajectory_filter.addMeasurement(detections, detections[0]->time_point);
-//     // }
-
-//     // RCLCPP_INFO(this->pnode->get_logger(), "IMG_CALLBACK DONE PROCESSING -- %s", (std::stringstream{} << std::this_thread::get_id()).str().c_str());
-
-//     // filter
-//     TagDetection::Ptr best_detection = nullptr;
-//     double best_score = 0.;
-//     {
-//         for(auto& detection : detections)
-//         {
-//             const double
-//                 tags_per_range = detection->num_tags / detection->avg_range,
-//                 rms_per_tag = detection->rms / detection->num_tags;
-//             const bool
-//                 in_bounds = this->pnode->tag_filtering.filter_bbox.isEmpty() ||
-//                     this->pnode->tag_filtering.filter_bbox.contains(detection->pose.vec),
-//                 tags_per_range_ok = tags_per_range >= this->pnode->tag_filtering.thresh_min_tags_per_range,
-//                 rms_per_tag_ok = rms_per_tag <= this->pnode->tag_filtering.thresh_max_rms_per_tag,
-//                 pix_area_ok = detection->pix_area >= this->pnode->tag_filtering.thresh_min_pix_area;
-
-//             if(in_bounds && tags_per_range_ok && rms_per_tag_ok && pix_area_ok)
-//             {
-//                 const double score = detection->rms;
-
-//                 // update best
-//                 if(!best_detection || score < best_score)
-//                 {
-//                     best_detection = detection;
-//                     best_score = score;
-//                 }
-//             }
-//             else
-//             {
-//                 detection = nullptr;
-//             }
-//         }
-//     }
-//     if(best_detection)
-//     {
-//         this->pnode->trajectory_filter.addMeasurement(best_detection, best_detection->time_point);
-//     }
-
-//     // TODO: handle initialization case where we don't have odometry yet
-//     // if(!this->pnode->state.dlo_in_progress && best_detection)
-//     // {
-//     //     // hard realign + publish tf
-
-//     //     this->pnode->state.tf_mtx.lock();
-
-//     //     this->pnode->state.map_tf.tf = (Eigen::Translation3d{ best_detection->translation } * best_detection->rotation) *
-//     //         (this->pnode->state.odom_tf.pose.quat.inverse() * Eigen::Translation3d{ -this->pnode->state.odom_tf.pose.vec });
-//     //     this->pnode->state.map_tf.pose << this->pnode->state.map_tf.tf;
-
-//     //     this->pnode->sendTf(img->header.stamp, false);
-
-//     //     this->pnode->state.tf_mtx.unlock();
-//     // }
-
-//     // cache detections for DLO refinement
-//     // this->pnode->state.alignment_mtx.lock();
-//     // for(const auto& detection : detections)
-//     // {
-//     //     if(detection) this->pnode->alignment_queue.push_front(detection);
-//     // }
-//     // this->pnode->state.alignment_mtx.unlock();
-//     // if(best_detection)
-//     // {
-//     //     // this->pnode->state.alignment_mtx.lock();
-//     //     // this->pnode->alignment_queue.push_front(best_detection);
-//     //     // this->pnode->state.alignment_mtx.unlock();
-//     //     // util::geom::Pose3d p;
-//     //     // p.vec = best_detection->translation;
-//     //     // p.quat = best_detection->rotation;
-
-//     //     this->pnode->trajectory_filter.addMeasurement(best_detection->pose, best_detection->time_point);
-//     // }
-
-//     auto _end = std::chrono::system_clock::now();
-//     const double _dt = this->pnode->metrics.img_thread.addSample(_start, _end);
-//     this->pnode->appendThreadProcTime(ProcType::IMG_CB, _dt);
-
-//     this->pnode->handleStatusUpdate();
-//     this->pnode->handleDebugFrame();
-
-//     // RCLCPP_INFO(this->pnode->get_logger(), "IMG_CALLBACK EXIT -- %s", (std::stringstream{} << std::this_thread::get_id()).str().c_str());
-// }
-
-// void PerceptionNode::CameraSubscriber::info_callback(const sensor_msgs::msg::CameraInfo::ConstSharedPtr& info)
-// {
-//     // RCLCPP_INFO(this->pnode->get_logger(), "INFO_CALLBACK");
-//     auto _start = std::chrono::system_clock::now();
-
-//     if( !this->valid_calib &&
-//         info->k.size() == 9 &&
-//         info->d.size() >= 5 )
-//     {
-//         this->calibration = cv::Mat(info->k, true).reshape(0, 3);
-//         this->distortion = cv::Mat(info->d, true).reshape(0, 1);
-
-//         this->valid_calib = true;
-
-//         // log
-//     }
-
-//     auto _end = std::chrono::system_clock::now();
-//     const double _dt = this->pnode->metrics.info_thread.addSample(_start, _end);
-//     this->pnode->appendThreadProcTime(ProcType::INFO_CB, _dt);
-
-//     this->pnode->handleStatusUpdate();
-//     this->pnode->handleDebugFrame();
-
-//     // RCLCPP_INFO(this->pnode->get_logger(), "INFO_CALLBACK EXIT");
-// }
 
 
 void PerceptionNode::getParams()
 {
-    // this->declare_parameter("image_transport", "raw");
-
     util::declare_param(this, "map_frame_id", this->map_frame, "map");
     util::declare_param(this, "odom_frame_id", this->odom_frame, "odom");
     util::declare_param(this, "base_frame_id", this->base_frame, "base_link");
 
     util::declare_param(this, "debug.status_max_print_freq", this->param.status_max_print_freq, 10.);
     util::declare_param(this, "debug.img_max_pub_freq", this->param.img_debug_max_pub_freq, 30.);
-
-    // std::vector<double> min, max;
-    // util::declare_param(this, "tag_filtering.bounds_min", min, {});
-    // util::declare_param(this, "tag_filtering.bounds_max", max, {});
-    // if(min.size() > 2 && max.size() > 2)
-    // {
-    //     this->tag_filtering.filter_bbox = Eigen::AlignedBox3d{ *reinterpret_cast<Eigen::Vector3d*>(min.data()), *reinterpret_cast<Eigen::Vector3d*>(max.data()) };
-    // }
-
-    // util::declare_param(this, "tag_filtering.fitness.oob_weight", this->tag_filtering.fitness_oob_weight, 100.);
-    // util::declare_param(this, "tag_filtering.fitness.rms_weight", this->tag_filtering.fitness_rms_weight, 10.0);
-    // util::declare_param(this, "tag_filtering.thresh.max_linear_diff_velocity", this->tag_filtering.thresh_max_linear_diff_velocity, 5.);
-    // util::declare_param(this, "tag_filtering.thresh.max_angular_diff_velocity", this->tag_filtering.thresh_max_angular_diff_velocity, 5.);
-    // util::declare_param(this, "tag_filtering.thresh.min_tags_per_range", this->tag_filtering.thresh_min_tags_per_range, 0.5);
-    // util::declare_param(this, "tag_filtering.thresh.max_rms_per_tag", this->tag_filtering.thresh_max_rms_per_tag, 0.1);
-    // util::declare_param(this, "tag_filtering.thresh.min_sum_pix_area", this->tag_filtering.thresh_min_pix_area, 10000.);
-
-    // util::declare_param(this, "tag_filtering.covariance.linear_base_coeff", this->tag_filtering.covariance_linear_base_coeff, 0.001);
-    // util::declare_param(this, "tag_filtering.covariance.linear_range_coeff", this->tag_filtering.covariance_linear_range_coeff, 0.001);
-    // util::declare_param(this, "tag_filtering.covariance.angular_base_coeff", this->tag_filtering.covariance_angular_base_coeff, 0.001);
-    // util::declare_param(this, "tag_filtering.covariance.angular_range_coeff", this->tag_filtering.covariance_angular_range_coeff, 0.001);
-    // util::declare_param(this, "tag_filtering.covariance.linear_rms_per_tag_coeff", this->tag_filtering.covariance_linear_rms_per_tag_coeff, 0.01);
-    // util::declare_param(this, "tag_filtering.covariance.angular_rms_per_tag_coeff", this->tag_filtering.covariance_angular_rms_per_tag_coeff, 0.01);
 }
 
 void PerceptionNode::initMetrics()
@@ -344,8 +125,6 @@ void PerceptionNode::initMetrics()
 void PerceptionNode::sendTf(const builtin_interfaces::msg::Time& stamp, bool needs_lock)
 {
     if(needs_lock) this->state.tf_mtx.lock();
-
-    // Eigen::Isometry3d full_tf = this->state.map_tf.pose.vec_trl() * (this->state.odom_tf.tf * this->state.map_tf.pose.quat);    // rotate then add translation
 
     geometry_msgs::msg::TransformStamped _tf;
     _tf.header.stamp = stamp;
@@ -434,19 +213,19 @@ void PerceptionNode::handleStatusUpdate()
                 << "+- CALLBACKS -------------------------------------------------------+\n"
                 << "|                        Comp. Time | Avg. Time | Max Time | Total  |\n";
             msg << std::setprecision(1) << std::fixed << std::right << std::setfill(' ');
-            this->metrics.info_thread.mtx.lock();
-            msg << "|  Info CB (" << std::setw(5) << 1. / this->metrics.info_thread.avg_call_delta
-                                 << " Hz) ::  " << std::setw(5) << this->metrics.info_thread.last_comp_time * 1e6
-                                               << " us  | " << std::setw(5) << this->metrics.info_thread.avg_comp_time * 1e6
-                                                           << " us  | " << std::setw(5) << this->metrics.info_thread.max_comp_time * 1e3
-                                                                       << " ms | " << std::setw(6) << this->metrics.info_thread.samples
-                                                                                   << " |\n";
-            this->metrics.info_thread.mtx.unlock();
+            // this->metrics.info_thread.mtx.lock();
+            // msg << "|  Info CB (" << std::setw(5) << 1. / this->metrics.info_thread.avg_call_delta
+            //                      << " Hz) ::  " << std::setw(5) << this->metrics.info_thread.last_comp_time * 1e6
+            //                                    << " us  | " << std::setw(5) << this->metrics.info_thread.avg_comp_time * 1e6
+            //                                                << " us  | " << std::setw(5) << this->metrics.info_thread.max_comp_time * 1e3
+            //                                                            << " ms | " << std::setw(6) << this->metrics.info_thread.samples
+            //                                                                        << " |\n";
+            // this->metrics.info_thread.mtx.unlock();
             this->metrics.img_thread.mtx.lock();
-            msg << "| Image CB (" << std::setw(5) << 1. / this->metrics.img_thread.avg_call_delta
-                                 << " Hz) ::  " << std::setw(5) << this->metrics.img_thread.last_comp_time * 1e3
-                                               << " ms  | " << std::setw(5) << this->metrics.img_thread.avg_comp_time * 1e3
-                                                           << " ms  | " << std::setw(5) << this->metrics.img_thread.max_comp_time * 1e3
+            msg << "|   DET CB (" << std::setw(5) << 1. / this->metrics.img_thread.avg_call_delta
+                                 << " Hz) ::  " << std::setw(5) << this->metrics.img_thread.last_comp_time * 1e6
+                                               << " us  | " << std::setw(5) << this->metrics.img_thread.avg_comp_time * 1e6
+                                                           << " us  | " << std::setw(5) << this->metrics.img_thread.max_comp_time * 1e3
                                                                        << " ms | " << std::setw(6) << this->metrics.img_thread.samples
                                                                                    << " |\n";
             this->metrics.img_thread.mtx.unlock();
@@ -474,11 +253,11 @@ void PerceptionNode::handleStatusUpdate()
             size_t idx = 0;
             for(auto& p : this->metrics.thread_proc_times)
             {
-                static constexpr char const* CHAR_VARS = "IiSuFMx"; // Img, info, Scan, imu, debug Frame, Metrics, misc
+                static constexpr char const* CHAR_VARS = "DSuFMx"; // Det, Scan, imu, debug Frame, Metrics, misc
                 static constexpr char const* COLORS[7] =
                 {
                     "\033[38;5;49m",
-                    "\033[38;5;11m",
+                    // "\033[38;5;11m",
                     "\033[38;5;45m",
                     "\033[38;5;198m",
                     "\033[38;5;228m",
@@ -520,89 +299,30 @@ void PerceptionNode::handleStatusUpdate()
     }
 }
 
-// void PerceptionNode::handleDebugFrame()
-// {
-//     // RCLCPP_INFO(this->get_logger(), "DEBUG FRAME INIT");
-//     // try lock mutex
-//     if(this->state.any_new_frames && this->state.frames_mtx.try_lock())
-//     {
-//         // check frequency
-//         auto _tp = std::chrono::system_clock::now();
-//         const double _dt = util::toFloatSeconds(_tp - this->state.last_frames_time);
-//         if(_dt > (1. / this->param.img_debug_max_pub_freq))
-//         {
-//             this->state.last_frames_time = _tp;
-//             this->state.any_new_frames = false;
 
-//             // RCLCPP_INFO(this->get_logger(), "DEBUG FRAME RUNNING...");
+void PerceptionNode::detection_callback(const cardinal_perception::msg::TagsDetection::ConstSharedPtr& detection_group)
+{
+    auto _start = std::chrono::system_clock::now();
 
-//             // lock & collect frames
-//             std::vector<cv::Mat> frames;
-//             std::vector<std::unique_lock<std::mutex>> locks;
-//             frames.reserve(this->camera_subs.size());
-//             locks.reserve(this->camera_subs.size());
-//             int max_height{ 0 };
+    const geometry_msgs::msg::TransformStamped& tf = detection_group->estimated_tf;
+    if(tf.header.frame_id == this->map_frame && tf.child_frame_id == this->base_frame)
+    {
+        TagDetection::Ptr td = std::make_shared<TagDetection>();
+        td->pose << tf.transform;
+        td->time_point = util::toFloatSeconds(tf.header.stamp);
+        td->pix_area = detection_group->pix_area;
+        td->avg_range = detection_group->avg_range;
+        td->rms = detection_group->rms;
+        td->num_tags = detection_group->num_tags;
+        this->trajectory_filter.addMeasurement(td, td->time_point);
+    }
 
-//             for(CameraSubscriber& s : this->camera_subs)
-//             {
-//                 std::unique_lock<std::mutex> _lock;
-//                 // cv::Mat& _frame = s.dbg_frame.B(_lock);
-//                 cv::Mat& _frame = s.dbg_frame.lock(_lock);
-//                 // s.dbg_frame.try_spin();     // hint a spin since we are done with all the current frames?
-//                 if(_frame.size().area() > 0)
-//                 {
-//                     if(_frame.size().height > max_height) max_height = _frame.size().height;
-//                     frames.push_back(_frame);
-//                     locks.emplace_back(std::move(_lock));
-//                 }
-//             }
+    auto _end = std::chrono::system_clock::now();
+    const double _dt = this->metrics.img_thread.addSample(_start, _end);
+    this->appendThreadProcTime(ProcType::DET_CB, _dt);
 
-//             for(size_t i = 0; i < frames.size(); i++)
-//             {
-//                 // resize frames, unlock on dereferences
-//                 if(frames[i].size().height < max_height)
-//                 {
-//                     double ratio = max_height / frames[i].size().height;
-//                     cv::Mat _resized;
-//                     cv::resize(frames[i], _resized, cv::Size{}, ratio, ratio, cv::INTER_AREA);
-//                     frames[i] = _resized;
-//                     locks[i].unlock();
-//                 }
-//             }
-
-//             // combine & output
-//             if(frames.size() > 0)
-//             {
-//                 try
-//                 {
-//                     thread_local cv::Mat pub;
-//                     if(frames.size() > 1)
-//                     {
-//                         cv::hconcat(frames, pub);
-//                     }
-//                     else
-//                     {
-//                         pub = frames[0];
-//                     }
-
-//                     std_msgs::msg::Header hdr;
-//                     hdr.frame_id = this->base_frame;
-//                     hdr.stamp = this->get_clock()->now();
-
-//                     this->debug_img_pub.publish(cv_bridge::CvImage(hdr, "bgr8", pub).toImageMsg());
-//                 }
-//                 catch(const std::exception& e)
-//                 {
-//                     // TODO
-//                 }
-//             }
-//         }
-//         this->appendThreadProcTime(ProcType::HANDLE_DBG_FRAME, util::toFloatSeconds(std::chrono::system_clock::now() - _tp));
-//         this->state.frames_mtx.unlock();
-//     }
-//     // RCLCPP_INFO(this->get_logger(), "DEBUG FRAME EXIT");
-// }
-
+    this->handleStatusUpdate();
+}
 
 void PerceptionNode::scan_callback(const sensor_msgs::msg::PointCloud2::ConstSharedPtr& scan)
 {
