+++ src/engine/localization.hpp
@@ -16,6 +16,8 @@
 #include "geometry.hpp"
 #include "trajectory_filter.hpp"
 
+#include "cardinal_perception/msg/tags_detection.hpp"
+
 #include <array>
 #include <deque>
 #include <mutex>
@@ -46,8 +48,6 @@
 
 #include <nav_msgs/msg/path.hpp>
 
-// #include <image_transport/image_transport.hpp>
-
 #include <boost/circular_buffer.hpp>
 
 #include <Eigen/Core>
@@ -58,9 +58,6 @@
 #include <pcl/filters/voxel_grid.h>
 #include <pcl/surface/concave_hull.h>
 #include <pcl/surface/convex_hull.h>
-
-// #include <opencv2/core.hpp>
-// #include <opencv2/aruco.hpp>
 
 #if USE_GTSAM_PGO > 0
 #include <gtsam/geometry/Rot3.h>
@@ -76,28 +73,11 @@
 #include <nano_gicp/nano_gicp.hpp>
 
 
-// struct TagDescription
-// {
-//     using Ptr = std::shared_ptr<TagDescription>;
-//     using ConstPtr = std::shared_ptr<const TagDescription>;
-
-//     std::array<cv::Point3f, 4>
-//         world_corners,
-//         rel_corners;
-
-//     Eigen::Vector3d translation;
-//     Eigen::Quaterniond rotation;
-//     Eigen::Vector4d plane;
-
-//     static Ptr fromRaw(const std::vector<double>& world_corner_pts);
-// };
 struct TagDetection
 {
     using Ptr = std::shared_ptr<TagDetection>;
     using ConstPtr = std::shared_ptr<const TagDetection>;
 
-    // Eigen::Vector3d translation;
-    // Eigen::Quaterniond rotation;
     util::geom::Pose3d pose;
 
     double time_point, pix_area, avg_range, rms;
@@ -113,36 +93,6 @@ public:
     ~PerceptionNode() = default;
 
 protected:
-    // class CameraSubscriber
-    // {
-    // public:
-    //     CameraSubscriber() = default;
-    //     // CameraSubscriber(PerceptionNode* inst, const std::string& img_topic, const std::string& info_topic);
-    //     CameraSubscriber(const CameraSubscriber& ref);
-    //     ~CameraSubscriber() = default;
-
-    //     void initialize(PerceptionNode* inst, const std::string& img_topic, const std::string& info_topic);
-
-    // public:
-    //     PerceptionNode* pnode = nullptr;
-
-    //     image_transport::Subscriber image_sub;
-    //     rclcpp::Subscription<sensor_msgs::msg::CameraInfo>::SharedPtr info_sub;
-
-    //     // SpinBuffer<cv::Mat> dbg_frame;
-    //     Synchronized<cv::Mat> dbg_frame;
-    //     cv::Mat1d calibration = cv::Mat1d::zeros(3, 3);
-    //     cv::Mat1d distortion = cv::Mat1d::zeros(1, 5);
-
-    //     std::atomic<bool> valid_calib = false;
-    //     // bool valid_calib = false;
-
-    // private:
-    //     void img_callback(const sensor_msgs::msg::Image::ConstSharedPtr& img);
-    //     void info_callback(const sensor_msgs::msg::CameraInfo::ConstSharedPtr& info);
-
-    // };
-
     class DLOdom
     {
     friend PerceptionNode;
@@ -336,31 +286,6 @@ protected:
 
     };
 
-    // class TagDetector
-    // {
-    // friend PerceptionNode;
-    // public:
-    //     TagDetector(PerceptionNode* inst);
-    //     ~TagDetector() = default;
-
-    // protected:
-    //     void getParams();
-
-    //     // template<bool enable_debug = true>
-    //     void processImg(
-    //         const sensor_msgs::msg::Image::ConstSharedPtr& img,
-    //         PerceptionNode::CameraSubscriber& sub,
-    //         std::vector<TagDetection::Ptr>& detections);
-
-    // private:
-    //     PerceptionNode* pnode;
-
-    //     std::unordered_map<int, TagDescription::ConstPtr> obj_tag_corners;
-    //     cv::Ptr<cv::aruco::Dictionary> aruco_dict;
-    //     cv::Ptr<cv::aruco::DetectorParameters> aruco_params;
-
-    // };
-
     void getParams();
     void initPGO();
     void initMetrics();
@@ -368,8 +293,8 @@ protected:
     void sendTf(const builtin_interfaces::msg::Time& stamp, bool needs_lock = false);
 
     void handleStatusUpdate();
-    // void handleDebugFrame();
 
+    void detection_callback(const cardinal_perception::msg::TagsDetection::ConstSharedPtr& det);
     void scan_callback(const sensor_msgs::msg::PointCloud2::ConstSharedPtr& scan);
     void imu_callback(const sensor_msgs::msg::Imu::SharedPtr imu);
 
@@ -378,21 +303,19 @@ private:
     tf2_ros::TransformListener tf_listener;
     tf2_ros::TransformBroadcaster tf_broadcaster;
 
-    // image_transport::ImageTransport img_transport;
-
     rclcpp::CallbackGroup::SharedPtr mt_callback_group;
+    rclcpp::Subscription<cardinal_perception::msg::TagsDetection>::SharedPtr detections_sub;
     rclcpp::Subscription<sensor_msgs::msg::PointCloud2>::SharedPtr scan_sub;
     rclcpp::Subscription<sensor_msgs::msg::Imu>::SharedPtr imu_sub;
-    // std::vector<CameraSubscriber> camera_subs;
 
-    // image_transport::Publisher debug_img_pub;
     rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr filtered_scan_pub;
+#if USE_GTSAM_PGO > 0
     rclcpp::Publisher<nav_msgs::msg::Path>::SharedPtr path_pub;
+#endif
     FloatPublisherMap metrics_pub;
     PublisherMap<geometry_msgs::msg::PoseStamped> pose_pub;
 
     DLOdom lidar_odom;
-    // TagDetector tag_detection;
     TrajectoryFilter<TagDetection> trajectory_filter;
 
 #if USE_GTSAM_PGO > 0
@@ -413,8 +336,6 @@ private:
     }
     pgo;
 #endif
-
-    // std::deque<TagDetection::Ptr> alignment_queue;
 
     std::string map_frame;
     std::string odom_frame;
@@ -440,27 +361,6 @@ private:
     }
     param;
 
-    // struct
-    // {
-    //     Eigen::AlignedBox3d filter_bbox;
-
-    //     double fitness_oob_weight;
-    //     double fitness_rms_weight;
-    //     double thresh_max_linear_diff_velocity;
-    //     double thresh_max_angular_diff_velocity;
-    //     double thresh_min_tags_per_range;
-    //     double thresh_max_rms_per_tag;
-    //     double thresh_min_pix_area;
-
-    //     double covariance_linear_base_coeff;
-    //     double covariance_linear_range_coeff;
-    //     double covariance_angular_base_coeff;
-    //     double covariance_angular_range_coeff;
-    //     double covariance_linear_rms_per_tag_coeff;
-    //     double covariance_angular_rms_per_tag_coeff;
-    // }
-    // tag_filtering;
-
     struct ThreadMetrics
     {
         std::chrono::system_clock::time_point last_call_time;
@@ -475,8 +375,7 @@ private:
 
     enum class ProcType : size_t
     {
-        IMG_CB = 0,
-        INFO_CB,
+        DET_CB = 0,
         SCAN_CB,
         IMU_CB,
         HANDLE_DBG_FRAME,
