+++ src/engine/todo.txt
@@ -1,1 +1,2 @@
 deadlock avoidance (try_for(...) on all mutex locks)
+composable nodes
